Updated fractal functions to avoid color banding

+++ src/fractal.h
@@ -3,4 +3,4 @@
 #include "settings/fractal_settings.h"
 
 /// @brief Given the fractal type, returns a function that computes that fractal
-std::function<int(double, double, const FractalSettings&)> get_fractal_function(FractalType);
+std::function<float(double, double, const FractalSettings&)> get_fractal_function(FractalType);
